refactor core into @frctl/core
