+++ packages/core/src/adapter.js
