+++ packages/core/src/array-stream.js
