+++ packages/core/src/entities/collection.js
