+++ packages/core/src/entities/entity.js
