+++ packages/core/src/entities/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    Collection: require('./collection'),
+    Entity: require('./entity'),
+    Source: require('./source'),
+};
