+++ packages/core/src/highlighter.js
