+++ packages/core/src/index.js
@@ -1,0 +1,15 @@
+module.exports = {
+    Adapter: require('./adapter'),
+    ArrayStream: require('./array-stream'),
+    data: require('./data'),
+    entities: require('./entities'),
+    fs: require('./fs'),
+    highlighter: require('./highlighter'),
+    Log: require('./log'),
+    markdown: require('./markdown'),
+    mixins: require('./mixins'),
+    PromiseStream: require('./promise-stream'),
+    resolver: require('./resolver'),
+    shell: require('./shell'),
+    utils: require('./utils'),
+};
