+++ packages/core/src/mixins/configurable.js
