+++ packages/core/src/mixins/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+    collection: require('./collection'),
+    configurable: require('./configurable'),
+    emitter: require('./emitter'),
+    entity: require('./entity'),
+    heritable: require('./heritable'),
+    mix: require('./mix'),
+    source: require('./source'),
+};
