+++ packages/core/src/resolver.js
