+++ packages/core/src/shell.js
