+++ packages/core/test/support/env.js
@@ -1,0 +1,1 @@
+process.env.NODE_ENV = 'test';
