+++ packages/fractal/package.json
@@ -15,9 +15,9 @@
   },
   "homepage": "https://github.com/frctl/fractal",
   "dependencies": {
-    "@allmarkedup/fang": "^1.0.0",
     "@frctl/handlebars": "^1.2.0",
     "@frctl/mandelbrot": "^1.4.0",
+    "@frctl/core": "^0.0.0",
     "anymatch": "^2.0.0",
     "bluebird": "^3.5.2",
     "browser-sync": "^2.26.4",
@@ -29,24 +29,17 @@
     "execa": "^4.0.1",
     "express": "^4.14.0",
     "fs-extra": "^9.0.0",
-    "globby": "^6.0.0",
     "gray-matter": "^2.0.2",
     "handlebars": "^4.0.13",
-    "highlight.js": "^9.5.0",
     "inquirer": "^1.1.2",
-    "istextorbinary": "^3.3.0",
     "js-yaml": "^3.13.1",
     "liftoff": "^2.3.0",
     "lodash": "^4.17.11",
     "log-update": "^1.0.2",
-    "marked": "^0.7.0",
     "mime": "^1.3.4",
-    "minimist": "^1.2.2",
-    "mixwith": "^0.1.1",
     "nunjucks": "^3.2.1",
     "path-to-regexp": "^1.5.3",
     "portscanner": "^1.0.0",
-    "readable-stream": "^2.1.4",
     "require-all": "^2.0.0",
     "semver": "^5.3.0",
     "throat": "^3.0.0",
@@ -57,7 +50,7 @@
   "devDependencies": {
     "chai": "^4.1.1",
     "mocha": "^5.0.3",
-    "mock-fs": "^4.6.0",
+    "nyc": "^15.0.1",
     "sinon": "^9.0.1"
   },
   "bin": {
