+++ packages/fractal/src/api/components/collection.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const EntityCollection = require('../../core/entities/collection');
+const EntityCollection = require('@frctl/core').entities.Collection;
 
 module.exports = class ComponentCollection extends EntityCollection {
     constructor(config, items, parent) {
