+++ packages/fractal/src/api/variants/variant.js
@@ -2,8 +2,8 @@
 
 const Path = require('path');
 const _ = require('lodash');
-const utils = require('../../core/utils');
-const Entity = require('../../core/entities/entity');
+const utils = require('@frctl/core').utils;
+const Entity = require('@frctl/core').entities.Entity;
 
 module.exports = class Variant extends Entity {
     constructor(config, view, resources, parent) {
