+++ packages/fractal/src/cli/commands/new.js
@@ -5,9 +5,9 @@ const Path = require('path');
 const Handlebars = require('handlebars');
 const inquirer = require('inquirer');
 const execa = require('execa');
-const shell = require('../../core/shell');
+const shell = require('@frctl/core').shell;
 const fs = require('fs-extra');
-const helpers = require('../../core/utils');
+const helpers = require('@frctl/core').utils;
 
 module.exports = {
     command: 'new <path>',
