+++ packages/fractal/src/web/builder.js
@@ -5,9 +5,9 @@ const anymatch = require('anymatch');
 const Path = require('path');
 const _ = require('lodash');
 const fs = require('fs-extra');
-const Log = require('../core/log');
-const mix = require('../core/mixins/mix');
-const Emitter = require('../core/mixins/emitter');
+const Log = require('@frctl/core').Log;
+const mix = require('@frctl/core').mixins.mix;
+const Emitter = require('@frctl/core').mixins.emitter;
 const throat = require('throat');
 
 module.exports = class Builder extends mix(Emitter) {
