+++ packages/fractal/src/web/web.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const _ = require('lodash');
-const utils = require('../core/utils');
-const mix = require('../core/mixins/mix');
-const Configurable = require('../core/mixins/configurable');
-const Emitter = require('../core/mixins/emitter');
+const utils = require('@frctl/core').utils;
+const mix = require('@frctl/core').mixins.mix;
+const Configurable = require('@frctl/core').mixins.configurable;
+const Emitter = require('@frctl/core').mixins.emitter;
 const Server = require('./server');
 const Builder = require('./builder');
 const Theme = require('./theme');
